requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
PySide6>=6.6.0
Pillow>=10.0.0
PyYAML>=6.0
qtawesome>=1.3.0
